Update helper1.py
Added a "filter volume" line in the audio process in "Helper1.py" at line 204.
This line permit you to increase the volume of the converted file if this one is too weak.
You can easily increase or decrease the volume with the valor of filter volume. The more high is the number, the more loud is the sound...
In my example, 6 was perfect for me, a sound high enough and no saturation on Flipper.